The first approach to implement multithreaded method "Sprite::Create()".

+++ IslandEngine/include/Game/CrusSprite.h
@@ -19,7 +19,7 @@ namespace isle {
 class Sprite final : public CrusObject {
 public:
 
-    static Sprite Create(std::shared_ptr<Texture> const &texture, uint16 number, Rect const &rect, Point const &pivot, float pixelsPerUnit = 100);
+    static Sprite Create(std::shared_ptr<Texture> const &texture, uint32 number, Rect const &rect, Point const &pivot, float pixelsPerUnit = 100);
 
     Texture const &textureSheet() const;
